requirements.txt
langchain-core==1.2.16
langchain-groq==1.1.2
langgraph==1.0.10
pandas==2.3.3
python-dotenv==1.2.2
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==4.1.1
gunicorn==25.3.0